:truck: move: ファイル名変更
関数名と同様のファイル名に変更する
